NUC-9220, NUC-9160
NUC-9220: Specify region instead of availability zone
NUC-9160: Validate region

Change --availability-zone to --region.  Don't pass the "placement"
argument to run_instances().  The default behavior is to launch the
instance in the connected region.

Get the current list of region names from AWS.  Verify that the user-
specified region is in this list.

Change-Id: Ice9b687c0ac17277946398add960ace8b22fc2d6