NUC-9303, NUC-9161, NUC-9299, NUC-9311
NUC-9303: solo-cli "brkt" needs to detect and raise errors
NUC-9161: Show snapshot progress
NUC-9299: error creating Bracket system log and Bracket system GRUB
snapshots
NUC-9311: Solo CLI flakiness

* Throw an exception when a snapshot goes into an error state.
* Log snapshot progress once a minute.
* Move the call to create_root_snapshot() inside the try block, so that
errors are handled the same way as with the rest of the process.
* Use create_snapshot() instead of copy_snapshot() to create the GRUB
and log volumes.  When the metavisor AMI doesn't have permissions set
correctly, copy_snapshot() would result in a snapshot in the error
state.
* Sleep before checking the state of a snapshot or instance.  This gives
AWS some time to propagate the RunInstance or CreateSnapshot change.
Not sure if this is a long term fix for NUC-9311, but it's a simple hack
that will improve reliability in the short term.
* Add the beginnings of a unit test suite that can be run with "python
setup.py test".

Change-Id: If21c8f6a5fc62af0e73aca59b42e443c753abbae