NUC-9368: brkt-cli waits until very end to fail with duplicate ami name
Add a nonce to the AMI name, to guarantee uniqueness.  Amazon requires
image names to be unique and no longer than 128 characters long.  The
new name is

<original name> (encrypted 787ace7a)

Trim the original name if the new AMI name would exceed 128 characters.
Replace sys.exit() with exit() so that we don't get a traceback in the
command output.

Change-Id: I5013ae336a883d07b881a93a7fda62235c79b9d2