Make iam.json policy more uniform
Per request of PM, this removes the tag conditionals for some
permissions. The reasoning behind this was to make the permissions
easier to understand for customers (as opposed to explaining why
some permissions restrict by tags and others do not) and also,
since the tag names are liable to be changed by us, it'd be
good to have users still work without required changes to policy.

Testing done:
- Update my IAM policy with the new document. Exported those AWS
credentials and successfully encrypted an image.

Change-Id: Ic54485332ece00bc86053ece595e54f728935cef