NUC-9120: Handle invalid AWS credentials
Handle NoAuthHandlerFound when connecting to AWS.  Reorganize the error
handling code so that all the exceptions we know about are handled in
main().  Suppress Boto logging unless the user passes the -v option, so
that we don't see Boto spew on auth errors.

Make the cleanup code more robust.  Run it inside the finally block, so
that it runs even if an exception was thrown.  Put each cleanup call to
AWS in its own try/except, so that cleanup errors don't override outer
exception handling.

Update the unit test code with the latest security group changes.

Change-Id: I65eccfbf1a7a9fe6b3c3494fabb1c2060a93e234