YETI-53: Wait until the resource is available before creating tags
Add a _safe_get_instance() function that waits for up to 10 seconds
until the AWS API returns the instance object.  Add a
_wait_for_security_group() function that does the same thing for a
security group.  Call _safe_get_instance() or _wait_for_security_group()
before tagging an instance or security group.

Log cleanup errors at the warn level, since they don't affect the result
of the script.  When cleaning up the security group, don't wait for the
instance to terminate if the call to terminate_instance() raised an
exception.

Add DescribeSecurityGroups to iam.json.  It's required for
_wait_for_security_group().

Change-Id: Ie449df416ebaee59e5ec95f9905172ee12bd503a