Delete EBS volumes created by the copier instance
Tag volumes attached to the snapshotter and encryptor instances.  After
instances have terminated, search for any volumes that were unexpectedly
orphaned by AWS and delete them.

Add a cleanup step for the snapshotter instance, so that it gets
terminated if something goes wrong while snapshotting the root volume.

Make the encryptor session id the same kind of shorter nonce that we use
to avoid name conflicts.

Change-Id: I39762db6c33f904a33ac8c65d75a2e58a8fc8209