Console output needs to be more specific about console log being available later
New log message when the AWS API does not return a console log for the
encryptor instance:

"Encryptor console output is not currently available.  Wait a minute and
check the console output for instance i-ebe542 in the EC2 Management
Console."

Add a unit test that verifies that we handle the case when console
output is not available.

Change-Id: Id3b679550a1683816a658292f4c5cf5773123c4b